+++ R/plotDiagnostics.R
@@ -17,7 +17,7 @@
 #' @export
 #'
 #' @examples
-#' x <- SLOPE(abalone$x, abalone$y, sigma = 2, diagnostics = TRUE)
+#' x <- SLOPE(abalone$x, abalone$y, alpha = 2, diagnostics = TRUE)
 #' plotDiagnostics(x)
 plotDiagnostics <- function(object,
                             ind = max(object$diagnostics$penalty),