+++ src/regularizationPath.h
@@ -7,14 +7,14 @@ using namespace arma;
 using namespace Rcpp;
 
 template <typename T>
-void regularizationPath(vec& sigma,
+void regularizationPath(vec& alpha,
                         vec& lambda,
-                        double& sigma_max,
+                        double& alpha_max,
                         const T& x,
                         const mat& y,
                         const rowvec& y_scale,
                         const std::string lambda_type,
-                        const std::string sigma_type,
+                        const std::string alpha_type,
                         const double lambda_min_ratio,
                         const double q,
                         const std::string family,
@@ -23,7 +23,7 @@ void regularizationPath(vec& sigma,
   const sword n = x.n_rows;
   const uword m = y.n_cols;
   const sword n_lambda = lambda.n_elem;
-  const uword n_sigma = sigma.n_elem;
+  const uword path_length = alpha.n_elem;
 
   if (lambda_type == "gaussian" || lambda_type == "bh") {
     lambda = regspace(1, n_lambda)*q/(2*n_lambda);
@@ -61,12 +61,12 @@ void regularizationPath(vec& sigma,
                              family,
                              intercept);
 
-  sigma_max =
+  alpha_max =
     (cumsum(sort(abs(lambda_max), "descending"))/cumsum(lambda)).max();
 
-  if (sigma_type == "auto") {
-    sigma = exp(linspace(log(sigma_max),
-                         log(sigma_max*lambda_min_ratio),
-                         n_sigma));
+  if (alpha_type == "auto") {
+    alpha = exp(linspace(log(alpha_max),
+                         log(alpha_max*lambda_min_ratio),
+                         path_length));
   }
 }