+++ tests/testthat/test-predictions.R
@@ -6,7 +6,7 @@ test_that("predictions work for all models", {
     x <- xy$x
     y <- xy$y
 
-    fit <- SLOPE(x, y, family = family, n_sigma = 5)
+    fit <- SLOPE(x, y, family = family, path_length = 5)
 
     for (type in c("link", "response", "class")) {
       if (type == "class" && family %in% c("gaussian", "poisson"))