+++ tests/testthat/test-training.R
@@ -10,7 +10,7 @@ test_that("model training works with trainSLOPE", {
                       number = 2,
                       q = c(0.1, 0.2),
                       repeats = 2,
-                      n_sigma = 2)
+                      path_length = 2)
     expect_s3_class(fit, "TrainedSLOPE")
   }
 })